+++ Changes.md
@@ -27,7 +27,7 @@ Build
 API
 ---
 
-- PlugAlgo : Added support in `setValueFromData()` for setting StringPlug values from StringVectorData.
+- PlugAlgo : Added support in `setValueFromData()` for setting StringPlug values from StringVectorData and StringVectorDataPlugValues from StringData.
 
 1.5.12.0 (relative to 1.5.11.0)
 ========